+++ src/maidsafe/vault/parameters.h
@@ -28,10 +28,9 @@ struct Parameters {
   // Max number of recent entries in account classes.
   static size_t max_recent_data_list_size;
 
-  static void set_file_element_count_limits(size_t max_file_element_count,
-                                            size_t min_file_element_count);
-  static size_t max_file_element_count();
-  static size_t min_file_element_count();
+  static void set_file_element_count_limits(int max_file_element_count, int min_file_element_count);
+  static int max_file_element_count();
+  static int min_file_element_count();
  private:
   Parameters();
   ~Parameters();
@@ -41,9 +40,9 @@ struct Parameters {
   Parameters& operator=(Parameters&&);
 
   // Max count of elements allowed in each account file
-  static size_t max_file_element_count_;
+  static int max_file_element_count_;
   // Min count of elements allowed in each account file
-  static size_t min_file_element_count_;
+  static int min_file_element_count_;
 };
 
 }  // namespace detail
